w3c-reports/12092014/mac-wk2/runned-tests/html/semantics/embedded-content/media-elements/track/track-element/cors/023-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
Actual	Expected
{
  "event": "error",
  "requests": []
}
{
  "event": "error",
  "requests": [
    {
      "cors": "http://localhost:8800",
      "cookie": "yes"
    }
  ]
}
Test id: a09dd37d-38e3-49d5-9aa3-e2f8129dfc02



FAIL track CORS: Use Credentials, not same-origin, no headers, redirects to same-origin, no headers assert_object_equals: expected property "0" missing(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
check_equal@http://localhost:8800/resources/testharness.js:384:24
check_equal@http://localhost:8800/resources/testharness.js:375:37
assert_object_equals@http://localhost:8800/resources/testharness.js:390:21
checkData@http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/common.js?pipe=sub:128:29
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
